# Introduce minimal WireSchema