… go/otel/grpc_test.go
… go/otel/manual_test.go
… go/pkg/schema/schema.go
… go/pkg/schema/schema_test.go
… go/pkg/schema/testdata/oteltef.wire.json
… go/pkg/schema/wireschema.go
… stefgen/generator/writer.go
… stefgen/templates/dicts.go.tmpl
… stefgen/templates/oneof.go.tmpl
… stefgen/templates/struct.go.tmpl
… stefgen/templates/writer.go.tmpl