This decouples human-readable Schema that is in JSON
format from WireSchema that is serialized in binary
format for efficiency. WireSchema contains only the
information that is necessary for wire compatibility
checks.

Resolves https://github.com/splunk/stef/issues/24
(Note that we still need struct names for checks).

Here are byte size comparisons between JSON and WIRE
for the Otel/TEF schema with "Metrics" root:

```
JSON:  2976, zstd:  760
WIRE:   115, zstd:  113
```

The following components were refactored to use WireSchema
instead of Schema:
- Generated Reader/Writer
- Encoder/Decoders
- gRPC client/server

The Generator continues using JSON Schema as its input, however
it now generates a WireSchema to be used in the generated code.